requirements.txt
rpaframework==23.0.0
pandas==2.0.1
